% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfifo.R
\name{add}
\alias{add}
\title{Add a value to the queue}
\usage{
add(q, val)
}
\arguments{
\item{q}{is the current fifo queue object}

\item{val}{is the value to be added to the queue}
}
\value{
The updated queue object
}
\description{
Add a value to the queue
}
\examples{
q <- qfifo()
q <- add(q,1234)
}
