% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfifo.R
\name{get_length}
\alias{get_length}
\title{Gets the length of the queue}
\usage{
get_length(q)
}
\arguments{
\item{q}{is the current fifo queue object}
}
\value{
The length of the queue
}
\description{
Gets the length of the queue
}
\examples{
q <- qfifo()
q <- add(q,1234)
q <- get_length(q)
}
