% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfifo.R
\name{process}
\alias{process}
\title{Delete the top element from teh queue}
\usage{
process(q)
}
\arguments{
\item{q}{is the current fifo queue object}
}
\value{
The modified queue
}
\description{
Delete the top element from teh queue
}
\examples{
q <- qfifo()
q <- add(q,1234)
q <- add(q,5678)
q <- process(q)
}
