% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfifo.R
\name{top}
\alias{top}
\title{Return the top value in the queue}
\usage{
top(q)
}
\arguments{
\item{q}{is the current fifo queue object}
}
\value{
The top of the queue
}
\description{
Return the top value in the queue
}
\examples{
q <- qfifo()
q <- add(q,1234)
v <- top(q)
}
